requirements.txt
mysql-connector-python==8.0.23
psycopg2-binary==2.8.6
sqlalchemy==1.4.1